support anchor copy
